First Initial commit. Note that this is still the rough version of the project so it need improvements.
I've started this project following these steps:

React
 1 create React project using 'npm create' command on visual studio code terminal
 2 install project dependencies using 'npm install'
 3 implement dummy ReviewList component to display reviews list
 4 fetch review data from the WordPress REST API
 5 parse and format the received review data
 6 render the reviews in a responsive list with basic styling

WordPress
 1 create wordPress plugin with data.json & PHP file
 2 configure XAMPP control panel with Apache & MySQL modules to host plugin locally
 3 save the WordPress plugin in XAMPP htdocs folder for local testing
 4 test the WordPress website API endpoints using Postman

This commit is an early stage of the project containing the initial implementation still in a rough state.
The React project is created using the 'npm create' command, & the ReviewList component is implemented to display a list of reviews.
The styling and layout of the review list are still pretty basic.

The data containing the reviw list, is fetched from the WordPress REST API, & formatted to extract relevant information.
The reviews are rendered dynamically in a responsive list, but the visual aspect is still in a rough stage (almost non existent).

The WordPress plugin is created with a PHP & json files to provide API endpoints for fetching reviews.
On my local machine, The XAMPP control panel is configured to host the WordPress plugin locally for testing and development.
I have used postman to first  test the functionality of the WordPress website API endpoints.