Second commit - logical implementation, more styling & project structure organization
	1 created a ReviewList.css which renders the visual aspect of the data on the browser
	2 installed the npm react-icons, this installation reaveled vulnarabilities that were fixed
	3 on the src directory I created an UI 'User Interface' folder that to contain files related to visual styling
	4 on the 'fullstack-exercise\src\UI' folder I created files to displays the visuals of each property of the Reviews extracted from the list component
	5 created a ReviewItem.js component that renders the necessary info for each review property extracted from the list
	6 created a LoadMoreBrands.js component, it has the code responsible for proper loading the brands in response to the user interaction and its "Load More" button
	7 the new implementations & restructuring of the projects structure is workinf properly so I aimed to render the visuals effects from the example given on the instructions pdf, but still needs improvements

this commit has all the React Application 'functional' aspects working. It also has some styling improvements, but still in a rough stage.
Part of the implementations of the ReviewList.js component were reorganized and passed to other components such as the ReviewItem.css.
The visual styling that the browser renders to the user has improved, such as the starRating icons corresponding to each integer of the "rating" property and the green check icons before each string of the "features" property
container classes were implemented for styling endings, and changes were made on each css item corresponding to the ReviewItem properties, such as font, color, margins, etc
added a container class for the play-now & terms classes to focus on there style and display them as a column

 /**
   * TODO overall styling & the visual aspects of the <div className="line"></div>
   * TODO add a visual style frame like to each Review item
   * TODO not sure if <div className="terms"> should display '21+ |' & '| Gamble Responsibly' or only the T&CS Apply with the anchor element
   */